refactor:AppModule軽量化
