+++ src/app/app-routing.module.ts
@@ -5,7 +5,6 @@ import { HeroComponent } from './heroes/hero/hero.component';
 import { HeroListComponent } from './heroes/hero-list/hero-list.component';
 import { VillainComponent } from './villains/villain/villain.component';
 import { VillainListComponent } from './villains/villain-list/villain-list.component';
-import { AuthGuard } from './shared/auth.guard';
 import { LoginComponent } from './login/login.component';
 
 const routes: Routes = [
@@ -23,7 +22,7 @@ const routes: Routes = [
   },
   {
     path: 'villains',
-    loadChildren: './villains/villains.mocule#VillainsModule',
+    loadChildren: './villains/villains.module#VillainsModule',
   },
 ];
