+++ src/app/app.module.ts
@@ -1,26 +1,21 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { LoginComponent } from './login/login.component';
-import { HeroesModule } from "app/heroes/heroes.module";
-import { VillainsModule } from "app/villains/villains.module";
 import { CoreModule } from "app/core/core.module";
 
 @NgModule({
   declarations: [
     AppComponent,
-    LoginComponent,
   ],
   imports: [
     BrowserModule,
-    HeroesModule,
-    VillainsModule,
     CoreModule,
     AppRoutingModule,
-    FormsModule,
+  ],
+  exports: [
+    CoreModule,
   ],
   providers: [ ],
   bootstrap: [AppComponent]
