+++ src/app/villains/villains-routing.module.ts
@@ -1,13 +1,13 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { AuthGuard } from "app/shared/auth.guard";
 import { VillainComponent } from "app/villains/villain/villain.component";
 import { VillainListComponent } from "app/villains/villain-list/villain-list.component";
+import { AuthGuard } from "app/core/auth.guard";
 
 const routes: Routes = [
     {
-    path: 'villains',
+    path: '',
     canActivate: [AuthGuard],    
     children: [
       {
